Rename header file to avoid name clash
