+++ src/DbConnection.h
@@ -3,7 +3,7 @@
 
 #include <boost/noncopyable.hpp>
 #include <boost/shared_ptr.hpp>
-#include "sqlite3.h"
+#include "sqlite3.hpp"
 
 class DbResult;
