+++ src/SqliteColumnDataSource.h
@@ -1,7 +1,7 @@
 #ifndef RSQLITE_SQLITECOLUMNDATASOURCE_H
 #define RSQLITE_SQLITECOLUMNDATASOURCE_H
 
-#include "sqlite3.h"
+#include "sqlite3.hpp"
 #include "DbColumnDataSource.h"
 
 class SqliteColumnDataSource : public DbColumnDataSource {
