+++ src/SqliteResultImpl.h
@@ -4,7 +4,7 @@
 
 #include <boost/noncopyable.hpp>
 #include <boost/shared_ptr.hpp>
-#include "sqlite3.h"
+#include "sqlite3.hpp"
 #include "DbColumnDataType.h"
 
 class DbConnection;
